refactor: Implement new project structure #2
