+++ app/domains/nanomag/endpoints.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 
 # Используем нашу общую утилиту
 from app.domains.common.utils import create_downloadable_response
-from database import get_db
+from app.database import get_db
 from .service import NanomagService
 from typing import Optional
